+++ database/migrations/2025_11_24_131303_create_opcaos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('id_var');
             $table->foreign('id_var')->references('id')->on('variacaos')->onDelete('cascade')->onUpdate('cascade');
             $table->string('nome');
-            $table->float('valor');
+            $table->decimal('valor',10,2);
             $table->timestamps();
         });
     }
